+++ lib/presentation/bloc/breeds_cubit.dart
@@ -2,7 +2,7 @@ import 'package:dogs/data/models/breeds_response_message.dart';
 import 'package:dogs/data/network/response/breed_images_response.dart';
 import 'package:dogs/data/network/response/breeds_response.dart';
 import 'package:dogs/data/repositories/breeds_repository.dart';
-import 'package:dogs/presentation/bloc/dogs_bloc/breeds_cubit_state.dart';
+import 'package:dogs/presentation/bloc/breeds_cubit_state.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class BreedsCubit extends Cubit<BreedsCubitState> {
