+++ lib/presentation/pages/fav_page.dart
@@ -1,5 +1,5 @@
-import 'package:dogs/presentation/bloc/app_data_provider.dart';
-import 'package:dogs/presentation/pages/breedimages/widgets/image_widget.dart';
+import 'package:dogs/presentation/widgets/image_widget.dart';
+import 'package:dogs/shared/providers/app_data_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
