+++ lib/presentation/widgets/image_widget.dart
@@ -1,5 +1,5 @@
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:dogs/presentation/bloc/app_data_provider.dart';
+import 'package:dogs/shared/providers/app_data_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
